feat: show statpopgen benchmarks
The diff is huge because I deduplicated a bunch of code between the web worker
and non-web-worker implementation.
