+++ benchmarks-website/index.html
@@ -261,6 +261,16 @@
                   "DuckDB:duckdb": "duckdb:duckdb",
                 },
               },
+           ],
+           [
+              "Statistical and Population Genetics",
+              {
+                renamedDatasets: {
+                  "DuckDB:vortex-file-compressed": "duckdb:vortex",
+                  "DuckDB:parquet": "duckdb:parquet",
+                  "DuckDB:duckdb": "duckdb:duckdb",
+                },
+              },
             ],
           ]);
         } else {
